PY/requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0

